backend/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pytesseract==0.3.13
Werkzeug==3.1.3
pdf2docx==0.5.5
pdf2image==1.16.3